backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
pydantic>=2.0,<3
httpx>=0.27,<1
SQLAlchemy>=2.0,<3
PyJWT>=2.8,<3
